Tweaked google VM to be able to run a mailserver on it.
Removed IPv4 address from the instance in order to be able to
create a PTR record for the instance. Since this works only for
IPv6 or static IPv4 addresses, I removed the ephemeral IPv4
address.

Also restructured the code a little bit to be more generic in
terms of the domain name.

Signed-off-by: Tim Beermann <tim@fintus.de>